src/requirements-dev.txt
mock==5.1.0
pika==1.3.2
Flask==3.0.3
pgcli==4.1.0
bandit==1.7.9
loguru==0.7.2
wheel==0.44.0
pytest==8.3.2
flake8==7.1.1
pymongo==4.8.0
Werkzeug==3.0.3
waitress==3.0.0
pip-tools==7.4.1
ggshield==1.30.2
requests==2.32.3
pytest-cov==5.0.0
sqlalchemy==2.0.32
setuptools==74.0.0
pytest-flask==1.3.0
pytest-mock==3.14.0
requests-mock==1.12.1
psycopg2-binary==2.9.9
flask_sqlalchemy==3.1.1
prometheus_client==0.20.0
python-dotenv<0.22.0,>=0.21.0
